requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.12" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
langchain-core==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
langchain-ollama==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
langgraph-checkpoint==1.0.10 ; python_version >= "3.12" and python_version < "4.0"
langgraph==0.2.22 ; python_version >= "3.12" and python_version < "4.0"
langsmith==0.1.121 ; python_version >= "3.12" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.1.1 ; python_version >= "3.12" and python_version < "4.0"
ollama==0.3.3 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.7 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.5 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
